Add files via upload
Day Cycle Script was used as a programming test script to work on the team as part of an interview. 

The remainder was made up of the first week working with the team, it's comprised of a Quest Object, a Task Object & it's children, a State Object & it's children, and a QuestManager where all quests were meant to be kept.